pyproject.toml
[project]
name = "football-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "langgraph>=0.2.0",
  "langchain-core>=0.3.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "rich>=13.7.0",

  # Providers (switchable)
  "langchain-openai>=0.2.0",
  "langchain-ollama>=0.2.0",
]

[tool.uv]
dev-dependencies = []